+++ client/src/components/DetailedRecipe/DetailedRecipe.css
@@ -1,14 +1,16 @@
 .DetailedRecipe__Container{
     display:grid;
     grid-template-columns: 1fr;
-    grid-template-rows: 1fr;
+    
 }
 
 
 
 .DetailedRecipe__Description{
     display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    justify-content: space-evenly;
+    grid-template-columns: 1fr 1fr;
+    
 }
