+++ client/src/components/DetailedRecipe/Instructions.css
@@ -1,37 +1,34 @@
-.Description__Ingredients{
-    display: grid;
-    grid-template-columns: 1fr;
-    margin-left:calc(30px + 20vmin);
+.Description__Instructions{
+    justify-self: center;
+    margin-right:calc(15px + 15vmin);
 }
 
-.Ingredients__Header{
+
+.Instructions--Heading{
+    font-weight:600;
+    /* margin-bottom:calc(2.5vmin); */
+}
+
+.Instructions__Wrapper{
+    column-count: 1;
+    column-width: 200px;
+    padding-left: 2em;
+}
+
+.Instruction--Content{
 
 }
 
-.Ingredients--Heading{
-  font-weight:600;
-  margin-bottom:calc(2.5vmin);
-}
-
-.Ingredients__List{
-  list-style-type: none;
-  padding-left: 2em;
-}
-
-.Ingredient--Item{
-  margin-top:calc(2.5vmin);
-  margin-bottom:calc(2.5vmin);
-}
 
 /*RESPONSIVE FONT BIG DESKTOP*/
 @media (min-width: 1441px) and (max-width: 2000px) {
-    .Ingredients--Heading{
+    .Instructions--Heading{
       font-size: calc(1.5rem + 16 * (100vw - 1441px) / (2000 - 1441) );
     }
 
-    .Ingredient--Item{
+    .Instruction--Content{
       font-size: calc(.5rem + 16 * (100vw - 1441px) / (2000 - 1441) );
-      line-height: calc( 120% + 4.2 * (100vw - 2000px) / (1441 - 2000) );
+      line-height: calc( 130% + 4.2 * (100vw - 2000px) / (1441 - 2000) );
     }
   }
 
@@ -39,13 +36,13 @@
 
 /*RESPONSIVE FONT DESKTOP*/
   @media (min-width: 800px) and (max-width: 1440px) {
-    .Ingredients--Heading{
+    .Instructions--Heading{
       font-size: calc(1.5rem + 16 * (100vw -800px) / (1440 - 800) );
     }
 
-    .Ingredient--Item{
+    .Instruction--Content{
       font-size: calc(.5rem + 16 * (100vw - 800px) / (1440 - 800) );
-      line-height: calc( 120% + 4.2 * (100vw - 2000px) / (1441 - 2000) );
+      line-height: calc( 130% + 4.2 * (100vw - 2000px) / (1441 - 2000) );
     }
 
   }
@@ -53,11 +50,11 @@
 
 /*RESPONSIVE FONT  MOBILE LANDSCAPE */
   @media (min-width: 640px) and (max-width: 1024px) and (min-height:360px) and (max-height:768px) {
-    .Ingredients--Heading{
+    .Instructions--Heading{
       font-size: calc(1.5rem + 16 * (100vw -640px) / (1024 - 640) );
     }
 
-    .Ingredient--Item{
+    .Instruction--Content{
       font-size: calc(.5rem + 16 * (100vw - 640px) / (1024 - 640) );
     }
 
@@ -65,11 +62,11 @@
 
   /* RESPONSIVE FONT  MOBILE VERTICAL*/
   @media (min-width: 320px) and (max-width: 500px) {
-    .Ingredients--Heading{
+    .Instructions--Heading{
       font-size: calc(1.5rem + 16 * (100vw -320px) / (500 - 320) );
     }
 
-    .Ingredient--Item{
+    .Instruction--Content{
       font-size: calc(.5rem + 16 * (100vw - 320px) / (500 - 320) );
     }
   }
